requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
requests
stripe
psycopg2-binary
gunicorn
geopy
Pillow
scikit-learn
pandas
numpy
matplotlib
seaborn
Flask-CORS
Flask-Migrate
Werkzeug
cryptography
bcrypt
python-dateutil
qrcode
redis
psutil
Flask-Login
python-telegram-bot
gevent
user-agents
geoip2
